Just updated the movement script with a fall limit
I have taken out the wall checks for now it can be added easily again if
we decide to do the wall sliding with code rather than physics
materials. The choice is to go through all of the platform shapes and
have a bound on the ones we want or have it on the player but then not
being able to control it for things like moving platforms. I will
implement the first of the two because it will give us more choice.