Started changing the levels into prefabs
The first three levels have been converted so that all it's parts are
now prefabs. Also I have added an ExitLevel script so that we can no go
to the next level. Chris you will need to add the functionality to the
game manager. Also I have added the LoadNextLevel function in the
GameManager script.